# Conflicts:
#	backend/main

+++ frontend/src/app/components/backoffice/backoffice.component.ts
@@ -1,8 +1,9 @@
 import {Component, OnInit, ViewChild} from '@angular/core';
 import {UserService} from '../../services/user.service';
-import {LocationService} from "../../services/location.service";
-import {UserService} from '../../services/user.service';
-import {LocationService} from "../../services/location.service";
+import {LocationService} from '../../services/location.service';
+import {FormBuilder, FormGroup, Validators} from '@angular/forms';
+import {TokenStorageService} from '../../services/token-storage.service';
+import {Router} from '@angular/router';
 
 interface User {
   ID: number;
@@ -24,53 +25,57 @@ interface Location {
   styleUrls: ['./backoffice.component.css']
 })
 
-
 export class BackofficeComponent implements OnInit {
 
   userElements: User[];
   userHeadElements = ['ID', 'Nome', 'Role', ''];
   locationElements: Location[];
-  locationHeadElements = ['ID', 'Nome', 'Latitude', 'Longitude', ''];
+  locationHeadElements = ['ID', 'Latitude', 'Longitude', 'Nome', ''];
   private errorMessage: any;
   hiddenElement: boolean;
   hiddenElement2: boolean;
+  checkBox: boolean;
+  userValidationForm: FormGroup;
+  locationValidationForm: FormGroup;
 
-  constructor(private userService: UserService, private locationService: LocationService) {
+  constructor(private userService: UserService, private locationService: LocationService, public fb: FormBuilder,
+              private tokenStorage: TokenStorageService, private router: Router) {
     this.hiddenElement = false;
     this.hiddenElement2 = true;
-
-export class BackofficeComponent implements OnInit {
-
-  userElements: User[];
-  userHeadElements = ['ID', 'Nome', 'Role', ''];
-  locationElements: Location[];
-  locationHeadElements = ['ID', 'Nome', 'Latitude', 'Longitude', ''];
-  private errorMessage: any;
-  hiddenElement: boolean;
-  hiddenElement2: boolean;
-
-  constructor(private userService: UserService, private locationService: LocationService) {
-    this.hiddenElement = false;
-    this.hiddenElement2 = true;
-
+    this.checkBox = false;
+    this.userValidationForm = fb.group({
+      username: [null, [Validators.required, Validators.email]],
+      password: [null, Validators.required],
+      isAdmin: [null],
+    });
+    this.locationValidationForm = fb.group({
+      name: [null, Validators.required],
+      latitude: [null, Validators.required],
+      longitude: [null, [Validators.required]]
+    });
   }
 
   ngOnInit(): void {
-    this.userService.getUsers().subscribe(data => {
-        this.userElements = data.data;
-      },
-      err => {
-        this.errorMessage = err.error.message;
-      }
-    );
-    this.locationService.getLocation().subscribe(data => {
-      this.locationElements = data.data;
-       this.elements = data.data;
-      },
-      err => {
-        this.errorMessage = err.error.message;
-      }
-    );
+    if (this.checkIfTokenIsValid() === true) {
+      this.userService.getUsers().subscribe(data => {
+          this.userElements = data.data;
+        },
+        err => {
+          this.errorMessage = err.error.message;
+        }
+      );
+      this.locationService.getLocation().subscribe(data => {
+          this.locationElements = data.data;
+        },
+        err => {
+          this.errorMessage = err.error.message;
+        }
+      );
+    }
+  }
+
+  reloadPage(): void {
+    window.location.reload();
   }
 
   onClickUsersHandler() {
@@ -83,28 +88,57 @@ export class BackofficeComponent implements OnInit {
     this.hiddenElement2 = false;
   }
 
-  onClickDeleteUser(id:number){
-    console.log("passou crlh")
+  onClickDeleteUser(id: number) {
+    if (this.checkIfTokenIsValid() === true) {
     this.userService.deleteUser(id).subscribe(data => {
-      this.userElements.forEach((user,index) =>{
-        if(user.ID===id){
-          this.userElements.splice(index,1)
-          
+      this.userElements.forEach((user, index) => {
+        if (user.ID === id) {
+          this.userElements.splice(index, 1);
         }
-      })
+      });
+    });
+    }
+  }
+
+  onClickDeleteLocation(id: number) {
+    if (this.checkIfTokenIsValid() === true) {
+      this.locationService.deleteLocation(id).subscribe(data => {
+        this.locationElements.forEach((location, index) => {
+          if (location.ID === id) {
+            this.locationElements.splice(index, 1);
+          }
+        });
+      });
+    }
+  }
+
+  onUserSubmit() {
+    if (this.checkIfTokenIsValid() === true) {
+      if (this.userValidationForm.value.isAdmin === true) {
+        this.userValidationForm.value.isAdmin = 'admin';
+      } else {
+        this.userValidationForm.value.isAdmin = 'user';
+      }
+      this.userService.addUser(this.userValidationForm.value).subscribe(data => {
+        console.log(data);
+        this.userElements.push(data.user);
+      });
+    }
+  }
+
+  onLocationSubmit() {
+    this.locationService.addLocation(this.locationValidationForm.value).subscribe(data => {
+      console.log(data);
+      this.locationElements.push(data.location);
     });
   }
 
-
-  onClickDeleteLocation(id:number){
-    console.log("passou crlh")
-    this.locationService.deleteLocation(id).subscribe(data => {
-      this.locationElements.forEach((location,index) =>{
-        if(location.ID===id){
-          this.locationElements.splice(index,1)
-          
-        }
-      })
-    });
+  checkIfTokenIsValid() {
+    if (!this.tokenStorage.getToken()) {
+      this.router.navigate(['/login']).then(r =>
+        this.reloadPage()
+      );
+    }
+    return true;
   }
 }
